Half implement CS-BuFLO (Incomplete).
What exists works but is missing the congestion sensitivity and dynamic
rate adaption (rho management) so it will attempt to spit frames out
onto the network every 1 ms unless the code thinks that it should be
idle.

Naturally this consumes bandwidth like it's popcorn, but all of the
"hard" parts of the code is the stuff required to get to this point, so
I count that as progress.