Add some quick and dirty stat reporting, and tune CS-BuFLO.
The parameters chosen for CS-BuFLO are based on the on-the-wire Tor
cell size and wild guessing.  This seems to bring the bandwidth
consumption down somewhat, though it's still not anything near what I
would personally call cheap.

The original pameters are still in basket.go for those that want to
experiment.